# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/chat": {
            "get": {
                "tags": [
                    "用户模块"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "userId1",
                        "name": "userId1",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "userId2",
                        "name": "userId2",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contact/addFriend": {
            "post": {
                "tags": [
                    "用户模块"
                ],
                "summary": "添加好友",
                "parameters": [
                    {
                        "type": "string",
                        "description": "发起请求的用户ID",
                        "name": "userId",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "目标用户ID",
                        "name": "targetId",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": -1, \"message\": \"添加失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contact/addMan": {
            "post": {
                "tags": [
                    "用户模块"
                ],
                "summary": "群主添加userId为群管理员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "操作者的ID",
                        "name": "operator",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "群组ID",
                        "name": "groupId",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "待添加管理员的ID",
                        "name": "userId",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contact/blockFriend": {
            "post": {
                "tags": [
                    "用户模块"
                ],
                "summary": "拉黑好友",
                "parameters": [
                    {
                        "type": "string",
                        "description": "发起请求的用户ID",
                        "name": "userId1",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "待删除的好友ID",
                        "name": "userId2",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": -1, \"message\": \"删除失败!\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contact/createGroup": {
            "post": {
                "tags": [
                    "用户模块"
                ],
                "summary": "创建群聊",
                "parameters": [
                    {
                        "type": "string",
                        "description": "群主用户ID",
                        "name": "ownerId",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "群组名称",
                        "name": "groupName",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": -1, \"message\": \"创建失败!\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contact/deMemberFromGroup": {
            "post": {
                "tags": [
                    "用户模块"
                ],
                "summary": "从群聊中删除成员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "操作者的ID",
                        "name": "operator",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "群组ID",
                        "name": "groupId",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "待删除成员的ID",
                        "name": "userId",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contact/deleteFriend": {
            "post": {
                "tags": [
                    "用户模块"
                ],
                "summary": "删除好友",
                "parameters": [
                    {
                        "type": "string",
                        "description": "发起请求的用户ID",
                        "name": "userId1",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "待删除的好友ID",
                        "name": "userId2",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": -1, \"message\": \"删除失败!\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contact/deleteGroup": {
            "post": {
                "tags": [
                    "用户模块"
                ],
                "summary": "群主解散群",
                "parameters": [
                    {
                        "type": "string",
                        "description": "操作者的ID",
                        "name": "operator",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "群组ID",
                        "name": "groupId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contact/friendsOnlineList": {
            "post": {
                "tags": [
                    "用户模块"
                ],
                "summary": "好友在线列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "userId",
                        "name": "userId",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contact/getContactList": {
            "get": {
                "tags": [
                    "用户模块"
                ],
                "summary": "获取所有关系",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contact/getGroupList": {
            "get": {
                "tags": [
                    "用户模块"
                ],
                "summary": "获取群的列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contact/groupsList": {
            "post": {
                "tags": [
                    "用户模块"
                ],
                "summary": "加载群聊列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userId",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": -1, \"message\": \"获取失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contact/joinGroup": {
            "post": {
                "tags": [
                    "用户模块"
                ],
                "summary": "加入群聊",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userId",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "群组ID",
                        "name": "groupId",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": -1, \"message\": \"加入失败!\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contact/outGroup": {
            "post": {
                "tags": [
                    "用户模块"
                ],
                "summary": "退群",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userId",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "群组ID",
                        "name": "groupId",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contact/searchUsersByGroupId": {
            "post": {
                "tags": [
                    "用户模块"
                ],
                "summary": "按照groupId 找到所有的群成员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "群组ID",
                        "name": "groupId",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contact/unRegister": {
            "post": {
                "tags": [
                    "用户模块"
                ],
                "summary": "注销账号",
                "parameters": [
                    {
                        "type": "string",
                        "description": "userId",
                        "name": "userId",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/friendsList": {
            "post": {
                "tags": [
                    "用户模块"
                ],
                "summary": "好友列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userId",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": -1, \"message\": \"获取列表失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/index": {
            "get": {
                "tags": [
                    "首页"
                ],
                "summary": "获取首页",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "get": {
                "tags": [
                    "登录"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/toRegister": {
            "post": {
                "tags": [
                    "首页"
                ],
                "summary": "跳转到注册页面",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/createUser": {
            "post": {
                "tags": [
                    "用户模块"
                ],
                "summary": "新增用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "确认密码",
                        "name": "Identity",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "电话号码",
                        "name": "phone",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": -1, \"message\": \"新增用户失败!\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/deleteUser": {
            "post": {
                "tags": [
                    "用户模块"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userId",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": -1, \"message\": \"删除用户失败!\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/findUserByPhoneAndPwd": {
            "post": {
                "tags": [
                    "用户模块"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "手机号码",
                        "name": "phone",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": -1, \"message\": \"登录失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/getUserList": {
            "get": {
                "tags": [
                    "用户模块"
                ],
                "summary": "获取所有用户",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/updateUser": {
            "post": {
                "tags": [
                    "用户模块"
                ],
                "summary": "修改昵称和密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userId",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": -1, \"message\": \"修改用户失败!\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}